Tornando o notebook organizado e didático + comentários
