+++ Solution.py
@@ -2,20 +2,22 @@ from mip import OptimizationStatus
 from SolutionType import SolutionType
 
 class Solution():
+    # Constructor
     def __init__(self, model):
-        self.model = model
-        self.solution = self.model.optimize()
-        if self.solution == OptimizationStatus.INFEASIBLE:
+        self.model = model                                                  # Modelo
+        self.solution = self.model.optimize()                               # Solução
+        if self.solution == OptimizationStatus.INFEASIBLE:                  # Verifica se a solução é inviável
             self.solutionType = SolutionType.INFEASIBLE
             self.solutionValue = None
             self.variablesValues = None
             self.variables = None
-        else:
-            self.solutionValue = model.objective_value
-            self.variablesValues = [i.x for i in list(self.model.vars)]
-            self.variables = list(self.model.vars)
-            self.defineSolutionType()
+        else:                                                               # Caso a solução seja viável:
+            self.solutionValue = model.objective_value                      # Armaneza o valor da solução
+            self.variablesValues = [i.x for i in list(self.model.vars)]     # Armazena os valores das variáveis
+            self.variables = list(self.model.vars)                          # Armazena as variáveis
+            self.defineSolutionType()                                       # Define o tipo da solução (binária ou não)
     
+    # Método que define o tipo da solução
     def defineSolutionType(self):
         for i in self.variablesValues:
             if i == 0 or i == 1:
@@ -24,6 +26,3 @@ class Solution():
                 self.solutionType = SolutionType.NOT_BINARY
                 return
         self.solutionType = SolutionType.BINARY
-    
-    def modelCopy(self):
-        pass
